Change create product methods. Add child form with details depend from category
Removed the category selection dialog when creating a product. Created a form with general settings for the product. You can choose an additional form with product details that depend on the category. When creating and updating a product, a message pops up about the success of the operation. After creating a product, you are redirected to the product editing page, but without the ability to change the category.