Add a disabled property to the product card
Added disabled property to the product card.
Disabled input property is needed to disable the transition by link - transition by clicking on the product name or on the image. For example, when creating a product, the appearance of the product card is displayed, which does not yet exist in the database, so clicking on a link that does not yet exist does not make sense.