Change upload-task component interface
The component receives the path and file, then loads the image into the folder with the specified id. If the file already exists, then appends an index to the end of the name.
It then returns an image object with information.
The component disappears after the specified time.

Type: refactor